+++ src/cloudflare.ts
@@ -4,6 +4,8 @@ import {
 } from "./components/InngestCommHandler";
 import { headerKeys, queryKeys } from "./helpers/consts";
 
+export const name = "cloudflare-pages";
+
 /**
  * In Cloudflare, serve and register any declared functions with Inngest, making
  * them available to be triggered by events.
@@ -12,7 +14,7 @@ import { headerKeys, queryKeys } from "./helpers/consts";
  */
 export const serve: ServeHandler = (nameOrInngest, fns, opts) => {
   const handler = new InngestCommHandler(
-    "cloudflare-pages",
+    name,
     nameOrInngest,
     fns,
     {