+++ src/components/Inngest.test.ts
@@ -1,9 +1,9 @@
-import { EventPayload } from "inngest";
+import { EventPayload } from "@local";
+import { eventKeyWarning } from "@local/components/Inngest";
+import { envKeys, headerKeys } from "@local/helpers/consts";
+import { IsAny } from "@local/helpers/types";
 import { assertType } from "type-plus";
-import { envKeys } from "../helpers/consts";
-import { IsAny } from "../helpers/types";
 import { createClient } from "../test/helpers";
-import { eventKeyWarning } from "./Inngest";
 
 const testEvent: EventPayload = {
   name: "test",
@@ -52,7 +52,7 @@ describe("instantiation", () => {
 
 describe("send", () => {
   describe("runtime", () => {
-    const originalEnvEventKey = process.env[envKeys.EventKey];
+    const originalProcessEnv = process.env;
     const originalFetch = global.fetch;
 
     beforeAll(() => {
@@ -70,18 +70,12 @@ describe("send", () => {
     beforeEach(() => {
       // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-unsafe-call, @typescript-eslint/no-explicit-any
       (global.fetch as any).mockClear();
-    });
-
-    afterEach(() => {
-      if (originalEnvEventKey) {
-        process.env[envKeys.EventKey] = originalEnvEventKey;
-      } else {
-        delete process.env[envKeys.EventKey];
-      }
+      process.env = { ...originalProcessEnv };
     });
 
     afterAll(() => {
       global.fetch = originalFetch;
+      process.env = originalProcessEnv;
     });
 
     test("should fail to send if event key not specified at instantiation", async () => {
@@ -150,6 +144,90 @@ describe("send", () => {
 
       await expect(inngest.send([])).resolves.toBeUndefined();
       expect(global.fetch).not.toHaveBeenCalled();
+    });
+
+    test("should send env:foo if explicitly set", async () => {
+      const inngest = createClient({
+        name: "test",
+        eventKey: testEventKey,
+        env: "foo",
+      });
+
+      await expect(inngest.send(testEvent)).resolves.toBeUndefined();
+      expect(global.fetch).toHaveBeenCalledWith(
+        expect.stringContaining(`/e/${testEventKey}`),
+        expect.objectContaining({
+          method: "POST",
+          // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
+          headers: expect.objectContaining({
+            [headerKeys.Environment]: "foo",
+          }),
+        })
+      );
+    });
+
+    test("should send env:foo if set in INNGEST_ENV", async () => {
+      process.env[envKeys.Environment] = "foo";
+
+      const inngest = createClient({
+        name: "test",
+        eventKey: testEventKey,
+      });
+
+      await expect(inngest.send(testEvent)).resolves.toBeUndefined();
+      expect(global.fetch).toHaveBeenCalledWith(
+        expect.stringContaining(`/e/${testEventKey}`),
+        expect.objectContaining({
+          method: "POST",
+          // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
+          headers: expect.objectContaining({
+            [headerKeys.Environment]: "foo",
+          }),
+        })
+      );
+    });
+
+    test("should send explicit env:foo over env var if set in both", async () => {
+      process.env[envKeys.Environment] = "bar";
+
+      const inngest = createClient({
+        name: "test",
+        eventKey: testEventKey,
+        env: "foo",
+      });
+
+      await expect(inngest.send(testEvent)).resolves.toBeUndefined();
+      expect(global.fetch).toHaveBeenCalledWith(
+        expect.stringContaining(`/e/${testEventKey}`),
+        expect.objectContaining({
+          method: "POST",
+          // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
+          headers: expect.objectContaining({
+            [headerKeys.Environment]: "foo",
+          }),
+        })
+      );
+    });
+
+    test("should send env:foo if set in platform env key", async () => {
+      process.env[envKeys.VercelBranch] = "foo";
+
+      const inngest = createClient({
+        name: "test",
+        eventKey: testEventKey,
+      });
+
+      await expect(inngest.send(testEvent)).resolves.toBeUndefined();
+      expect(global.fetch).toHaveBeenCalledWith(
+        expect.stringContaining(`/e/${testEventKey}`),
+        expect.objectContaining({
+          method: "POST",
+          // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
+          headers: expect.objectContaining({
+            [headerKeys.Environment]: "foo",
+          }),
+        })
+      );
     });
   });