+++ src/components/Inngest.ts
@@ -1,6 +1,11 @@
 import { envKeys } from "../helpers/consts";
 import { devServerAvailable, devServerUrl } from "../helpers/devserver";
-import { devServerHost, isProd, processEnv } from "../helpers/env";
+import {
+  devServerHost,
+  inngestHeaders,
+  isProd,
+  processEnv,
+} from "../helpers/env";
 import type {
   PartialK,
   SendEventPayload,
@@ -11,12 +16,12 @@ import type {
   ClientOptions,
   EventNameFromTrigger,
   EventPayload,
+  FailureEventArgs,
   FunctionOptions,
   Handler,
   ShimmedFns,
   TriggerOptions,
 } from "../types";
-import { version } from "../version";
 import { InngestFunction } from "./InngestFunction";
 
 /**
@@ -111,6 +116,7 @@ export class Inngest<
     eventKey,
     inngestBaseUrl = "https://inn.gs/",
     fetch,
+    env,
   }: ClientOptions) {
     if (!name) {
       throw new Error("A name must be passed to create an Inngest instance.");
@@ -124,10 +130,9 @@ export class Inngest<
       console.warn(eventKeyWarning);
     }
 
-    this.headers = {
-      "Content-Type": "application/json",
-      "User-Agent": `InngestJS v${version}`,
-    };
+    this.headers = inngestHeaders({
+      inngestEnv: env,
+    });
 
     this.fetch = Inngest.parseFetch(fetch);
   }
@@ -386,17 +391,19 @@ export class Inngest<
           fns?: TFns;
 
           /**
-           * Leaving commented out; the feature can be added in a small PR
-           * when ready.
+           * Provide a function to be called if your function fails, meaning
+           * that it ran out of retries and was unable to complete successfully.
            *
-           * TODO Add user-facing comments here.
+           * This is useful for sending warning notifications or cleaning up
+           * after a failure and supports all the same functionality as a
+           * regular handler.
            */
-          // onFailure?: Handler<
-          //   Events,
-          //   TTriggerName,
-          //   TShimmedFns,
-          //   FailureEventArgs<Events[TTriggerName]>
-          // >;
+          onFailure?: Handler<
+            Events,
+            TTriggerName,
+            TShimmedFns,
+            FailureEventArgs<Events[TTriggerName]>
+          >;
         }),
     trigger: TTrigger,
     handler: Handler<Events, TTriggerName, TShimmedFns>