+++ src/next.test.ts
@@ -1,4 +1,4 @@
-import * as NextHandler from "./next";
+import * as NextHandler from "@local/next";
 import { testFramework } from "./test/helpers";
 
 testFramework("Next", NextHandler);